+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 - Enh: Added `transition/embed/notification-overview`
 - Cgh #1041: Changed @background-color-page to #f0f2f5'
 - Cgn #1232: Darken space description in the Space directory
+- Enh #1260: Space stream page to be embedded
 
 1.7.1 (Nov 27, 2023)
 --------------------